Move EvaluationIndex to debug package
(Not a 100% correct but somewhat more pleasing)
