+++ core/test/transformer/testContext.ts
@@ -1,5 +1,6 @@
+import Globals from 'src/Globals';
 import TransformContext from 'src/transformer/TransformContext';
 
 export function emptyContext(): TransformContext {
-  return new TransformContext(new Map(), []);
+  return new TransformContext(new Map(), [], new Globals());
 }
